Fixing a major type error in get_top_cut()
git-svn-id: svn://innoue/tracking/trunk@291 4cb6fdf0-3dc9-4801-8ffc-c6659000aa44

+++ iden/include/params1.h
@@ -94,7 +94,7 @@ public:
   {return pctle_threshold_;}
   int get_mask_radius() const
   {return mask_radius_;}
-  int get_top_cut() const
+  float get_top_cut() const
   {return top_cut_;}
